refactor: merged diagram and file module into one
