+++ package.json
@@ -26,7 +26,7 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "~9.1.11",
-    "@angular/cdk": "^9.2.4",
+    "@angular/cdk": "~9.2.4",
     "@angular/common": "~9.1.11",
     "@angular/compiler": "~9.1.11",
     "@angular/core": "~9.1.11",
@@ -38,14 +38,14 @@
     "@angular/router": "~9.1.11",
     "@commitlint/cli": "~11.0.0",
     "@commitlint/config-angular": "~11.0.0",
-    "@nestjs/common": "^7.6.13",
-    "@nestjs/swagger": "^4.7.15",
+    "@nestjs/common": "^7.6.14",
+    "@nestjs/swagger": "^4.8.0",
     "@ngx-translate/core": "~12.1.2",
     "@ngx-translate/http-loader": "~5.0.0",
-    "ngx-file-drop": "~9.0.1",
-    "husky": "~4.3.0",
+    "husky": "^4.3.8",
     "js-base64": "~3.6.0",
-    "lodash": "~4.17.20",
+    "lodash": "^4.17.21",
+    "ngx-file-drop": "~9.0.1",
     "ngx-filesize": "^2.0.16",
     "ngx-highlightjs": "~4.1.2",
     "ngx-permissions": "^8.1.1",
@@ -56,13 +56,13 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.9",
-    "@angular/cli": "~9.1.9",
+    "@angular-devkit/build-angular": "^0.901.15",
+    "@angular/cli": "^9.1.15",
     "@angular/compiler-cli": "~9.1.11",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/lodash": "~4.14.108",
-    "@types/node": "^12.11.1",
+    "@types/lodash": "^4.14.168",
+    "@types/node": "^12.20.6",
     "codelyzer": "^5.1.2",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
