+++ src/app/app-routing.module.ts
@@ -13,11 +13,7 @@ const routes: Routes = [
     loadChildren: () => import('./landing/landing.module').then(m => m.LandingModule),
   },
   {
-    path: 'files',
-    loadChildren: () => import('./file/file.module').then(m => m.FileModule),
-  },
-  {
-    path: 'diagram',
+    path: 'app',
     loadChildren: () => import('./diagram/diagram.module').then(m => m.DiagramModule),
   },
 ];
