+++ src/app/file/file.module.ts
@@ -2,13 +2,13 @@ import { NgModule } from '@angular/core';
 
 import { SharedModule } from '../shared/shared.module';
 
-import { UploadFileComponent } from './components/upload-file/upload-file.component';
-import { UserFilesComponent } from './components/user-files/user-files.component';
+// import { UploadFileComponent } from './components/upload-file/upload-file.component';
+// import { UserFilesComponent } from './components/user-files/user-files.component';
 import { FileRoutingModule } from './file-routing.module';
 
 const components = [
-  UploadFileComponent,
-  UserFilesComponent,
+  // UploadFileComponent,
+  // UserFilesComponent,
 ];
 
 @NgModule({
